Move methods from BestFirstSearch object to TreeOps
